refactor: use es6 compatible modules where possible (#1299)
