+++ package.json
@@ -46,13 +46,13 @@
     "@ngneat/until-destroy": "^8.1.4",
     "@sentry/browser": "^6.19.7",
     "angulartics2": "^10.1.0",
-    "crypto-js": "^4.1.1",
+    "crypto-es": "^1.2.7",
     "deep-object-diff": "^1.1.7",
     "faker": "^5.5.3",
-    "flag-icon-css": "^3.5.0",
+    "flag-icons": "^6.5.1",
     "hammerjs": "^2.0.8",
     "json-query": "^2.2.2",
-    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
     "md5": "^2.3.0",
     "moment": "^2.29.3",
     "ngx-markdown": "^11.2.0",
@@ -72,19 +72,18 @@
     "@angular/compiler-cli": "^11.2.4",
     "@babel/core": "^7.13.8",
     "@compodoc/compodoc": "1.1.11",
-    "@cypress/schematic": "^1.5.1",
+    "@cypress/schematic": "~1.5.1",
     "@schematics/angular": "^11.2.14",
     "@storybook/addon-actions": "^6.4.19",
     "@storybook/addon-essentials": "^6.4.19",
     "@storybook/addon-links": "6.4.19",
     "@storybook/angular": "^6.4.19",
-    "@types/crypto-js": "^4.0.2",
     "@types/faker": "^5.5.7",
     "@types/file-saver": "^2.0.3",
     "@types/hammerjs": "^2.0.41",
     "@types/jasmine": "~3.8.2",
     "@types/json-query": "^2.2.3",
-    "@types/lodash": "^4.14.168",
+    "@types/lodash-es": "^4.17.6",
     "@types/md5": "^2.3.2",
     "@types/node": "^14.14.31",
     "@types/pouchdb": "^6.4.0",
