+++ src/app/core/webdav/webdav-wrapper.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import webdav, { WebDAVClient, WebDAVClientOptions } from "webdav";
+import { createClient, WebDAVClient, WebDAVClientOptions } from "webdav/web";
 
 /**
  * This Class only wraps the webdav library module in order for it to be mock-able in tests
@@ -12,6 +12,6 @@ export class WebdavWrapperService {
     remoteUrl: string,
     options?: WebDAVClientOptions
   ): WebDAVClient {
-    return webdav.createClient(remoteUrl, options);
+    return createClient(remoteUrl, options);
   }
 }
